styling: make additional debt token alert less wordy
