--- apps/main/src/llamalend/features/manage-soft-liquidation/ui/alerts/AlertAdditionalCrvUsd.tsx
@@ -1,45 +1,0 @@
-import Alert from '@mui/material/Alert'
-import AlertTitle from '@mui/material/AlertTitle'
-import Stack from '@mui/material/Stack'
-import Typography from '@mui/material/Typography'
-import type { Decimal } from '@primitives/decimal.utils'
-import { t, Trans } from '@ui-kit/lib/i18n'
-import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { formatTokens } from '../action-infos/util'
-
-const { Spacing } = SizesAndSpaces
-
-type Props = {
-  debtTokenSymbol: string | undefined // this can be undefined if the userState query failed or is pending
-  missing: Decimal
-  balance: Decimal
-}
-
-export const AlertAdditionalCrvUsd = ({ missing, debtTokenSymbol: debtToken = t`debt tokens`, balance }: Props) => (
-  <Alert severity="error" variant="outlined">
-    <AlertTitle>{t`Additional ${debtToken} required`}</AlertTitle>
-
-    <Stack gap={Spacing.sm}>
-      <Typography variant="bodySRegular" color="textSecondary">
-        <Trans>
-          Your position cannot be closed because the outstanding debt amount is greater than the available collateral
-          value. Additionally, your wallet does not contain sufficient <strong>{debtToken}</strong> to cover the
-          remaining debt balance that exceeds your collateral.
-        </Trans>
-      </Typography>
-
-      <Typography variant="bodySRegular" color="textSecondary">
-        <Trans>
-          A minimum balance of <strong>{formatTokens({ symbol: debtToken, amount: missing })}</strong> is required in
-          your wallet to close the position.
-        </Trans>
-      </Typography>
-
-      <Typography variant="bodyXsRegular" color="textTertiary" component="span">
-        <Trans>
-          (You currently have <strong>{formatTokens({ symbol: debtToken, amount: balance })}</strong> in your wallet)
-        </Trans>
-      </Typography>
-    </Stack>
-  </Alert>
-)
