+++ packages/curve-ui-kit/src/shared/ui/InlineLink.tsx
@@ -7,8 +7,18 @@ import { ArrowTopRightIcon } from '../icons/ArrowTopRightIcon'
  * InlineLink component is a link that is displayed inline with text.
  * Tanstack Link manages both internal and external links
  */
-export const InlineLink = ({ to, children }: { to?: string; children: ReactNode }) => {
-  const isExternal = to?.startsWith('http')
+export const InlineLink = ({
+  to,
+  external = false,
+  hideIcon = false,
+  children,
+}: {
+  to?: string
+  external?: boolean
+  hideIcon?: boolean
+  children: ReactNode
+}) => {
+  const isExternal = external || to?.startsWith('http')
   return (
     <Link
       sx={{ color: 'currentColor', '&:hover': { textDecoration: 'none' } }}
@@ -17,7 +27,7 @@ export const InlineLink = ({ to, children }: { to?: string; children: ReactNode 
       {...(isExternal && { target: '_blank', rel: 'noreferrer noopener' })}
     >
       {children}
-      {isExternal && <ArrowTopRightIcon fontSize="small" sx={{ verticalAlign: 'bottom' }} />}
+      {isExternal && !hideIcon && <ArrowTopRightIcon fontSize="small" sx={{ verticalAlign: 'bottom' }} />}
     </Link>
   )
 }
